% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{add_rcitation}
\alias{add_rcitation}
\title{Add citation for data preparation}
\usage{
add_rcitation(mydoc, prep = "Brian Gulbis")
}
\arguments{
\item{mydoc}{A docx object}

\item{prep}{An optional character string with the name of the person who
prepared the data analysis}
}
\value{
A docx object
}
\description{
\code{add_rcitation} add a citation for R to a Microsoft Word document
}
\details{
This function takes a docx object and adds paragraphs which cite R and the
name of the person who prepared the data analysis.
}

