% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{create_tableone}
\alias{create_tableone}
\title{Create a results table using tableone}
\usage{
create_tableone(test, group = NULL, ident = "pie.id")
}
\arguments{
\item{test}{A data frame with the data to be summarized}

\item{group}{An optional character character vector indicating grouping
column(s); default is no groups}

\item{ident}{An optional character string with the name of the subject
identifier column}
}
\value{
A TableOne object
}
\description{
\code{create_tableone} make a table with results to be inserted into
Microsoft Word
}
\details{
This function takes a data frame and converts it to a TableOne object.
}
\seealso{
\code{\link[tableone]{CreateTableOne}}
}

