% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd_lookup.R
\name{find_icd_codes}
\alias{find_icd_codes}
\title{Lookup ICD9/10 codes by description}
\usage{
find_icd_codes(pattern, icd10 = FALSE, procedure = FALSE)
}
\arguments{
\item{pattern}{A regular expression}

\item{icd10}{A logical indicating whether to use ICD-10 codes (default) or
ICD-9 codes}

\item{procedure}{A logical indicating whether to use diagnosis codes
(default) or procedure codes}
}
\value{
A data frame
}
\description{
\code{find_icd_codes} takes a regular expression and returns a data frame
with matching ICD9/10 codes
}
\details{
This function takes a regular expression and matches it to the ICD9/10
description. A data frame containing the matching ICD9/10 codes is returned.
}

