% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd_lookup.R
\name{icd_description}
\alias{icd_description}
\title{Lookup ICD9/10 code description}
\usage{
icd_description(codes, icd10 = FALSE, procedure = FALSE)
}
\arguments{
\item{codes}{A character vector of ICD9/10 codes}

\item{icd10}{A logical indicating whether to use ICD-10 codes (default) or
ICD-9 codes}

\item{procedure}{A logical indicating whether to use diagnosis codes
(default) or procedure codes}
}
\value{
A data frame with columns: icd9.code and icd9.description
}
\description{
\code{icd_description} takes a vector of ICD codes and returns a data frame
with the corresponding description for each code
}
\details{
This function takes a character vector of ICD codes and returns a data frame
with two columns: icd.code and icd.description.
}

