% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd_lookup.R
\name{icd_lookup}
\alias{icd_lookup}
\title{Lookup ICD9/10 codes from CCS codes}
\usage{
icd_lookup(df, icd10 = FALSE, procedure = FALSE)
}
\arguments{
\item{df}{A data frame with columns: disease.state, type, code}

\item{icd10}{A logical indicating whether to use ICD-10 codes (default) or
ICD-9 codes}

\item{procedure}{A logical indicating whether to use diagnosis codes
(default) or procedure codes}
}
\value{
A data frame
}
\description{
\code{icd_lookup} takes a data frame with CCS codes and returns the
corresponding ICD9 codes
}
\details{
This function takes a data frame with three columns: disease.state, type, and
code. The column \code{disease.state} is a character field with the name of a
disease state which will be used for grouping. The column \code{type} is a
character field with either "ICD" or "CCS", which indicates the type of code.
The column \code{code} is a character field with the ICD or CCS code. For all
rows with CCS codes, the function will look-up the corresponding ICD9/10 code
and then return a data frame with two columns: disease.state and icd.code.
The procedure parameter is used to specify whether diagnosis codes or
procedure codes should be returned.
}

