% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{result_docx}
\alias{result_docx}
\title{Create a results document}
\usage{
result_docx(project, authors, template = NULL, title.names = NULL,
  title.style = "TitleDoc", bookmark = "start")
}
\arguments{
\item{project}{A character string with the name of the project}

\item{authors}{A character string with the name of the authors}

\item{template}{An optional character string with the filename of the docx
file to be used as a template}

\item{title.names}{An optional character vector with names of title styles
used}

\item{title.style}{An optional character string with the name of the title
style to use for project title, defaults to TitleDoc}

\item{bookmark}{An optional character string with the name of the bookmark in
the Word document template where the report should start, defaults to
"start"}
}
\value{
A docx object
}
\description{
\code{result_docx} make a docx object to insert results into
}
\details{
This function creates a docx object and inserts project title, authors, and
report date into the object. The docx object can then be written to a
Microsoft Word document.
}
\seealso{
\code{\link[ReporteRs]{docx}}
}

