% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{result_plot}
\alias{result_plot}
\title{Add a ggplot2 graph to docx object}
\usage{
result_plot(mydoc, graph, title = NULL)
}
\arguments{
\item{mydoc}{A docx object}

\item{graph}{A ggplot2 object}

\item{title}{An optional character string, if not NULL (default), the title
will be added above the plot}
}
\value{
A docx object
}
\description{
\code{result_plot} add a ggplot2 graph to the docx object which can then be
inserted into Microsoft Word
}
\details{
This function takes a ggplot2 graph and adds the plot to the docx object
which is returned. The docx object can then be written to a Microsoft Word
document.
}
\seealso{
\code{\link[ReporteRs]{addPlot}}
}

