% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{result_regrmod}
\alias{result_regrmod}
\title{Create a result table for regression model}
\usage{
result_regrmod(mydoc, mod, table.title, exp = TRUE)
}
\arguments{
\item{mydoc}{A docx object}

\item{mod}{A regression model}

\item{table.title}{A character string}

\item{exp}{An optional logical, passed to ShowRegTable}
}
\value{
A docx object
}
\description{
\code{result_regrmod} make a table with results of a regression model to be
inserted into Microsoft Word
}
\details{
This function takes a regression model and saves the results in a FlexTable,
which is then added to the docx object which is returned. The docx object can
then be written to a Microsoft Word document.
}
\seealso{
\code{\link[tableone]{ShowRegTable}}
}

