% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{result_table2}
\alias{result_table2}
\title{Create multiple result tables separated by a factor}
\usage{
result_table2(mydoc, test, split.by, table.title, group = "group")
}
\arguments{
\item{mydoc}{A docx object}

\item{test}{A data frame}

\item{split.by}{A character string indicating the name of the column used to
separate the data into multiple data frames}

\item{table.title}{A character string, will proceed the name of the factor}

\item{group}{An optional character string indicating the name of the column
to group on; defaults to "group", set group to NULL to remove grouping}
}
\value{
A docx object
}
\description{
\code{result_table2} make multiple tables with results to be inserted into
Microsoft Word
}
\details{
This function takes a data frame and converts it into a separate FlexTable
for each factor of a variable. The FlexTables are then added to the docx
object which is returned. The docx object can then be written to a Microsoft
Word document. Similar to \code{\link{result_table}} except the data is first
separated into multiple data frames based on the split.by column.
}
\seealso{
\code{\link[ReporteRs]{FlexTable}}
}

