% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_data.R
\name{summarize_labs}
\alias{summarize_labs}
\title{Summary calculations for lab data}
\usage{
summarize_labs(lab.data, units = "hours")
}
\arguments{
\item{lab.data}{A data frame with lab data}

\item{units}{An optional character string specifying the time units to use in
calculations, default is hours}
}
\value{
A data frame
}
\description{
\code{summarize_labs} summarizes lab data
}
\details{
This function takes a data frame with numeric lab result data and produces a
data frame with summary data for each patient and lab. The calculations
include: first, last, median, maximum, minimum, AUC, and time-weighted
average result.
}

