% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_data.R
\name{tidy_labs}
\alias{tidy_labs}
\title{Tidy lab results}
\usage{
tidy_labs(raw.data, censor = TRUE)
}
\arguments{
\item{raw.data}{A data frame with all scheduled medications}

\item{censor}{A logical, will check for censored data if TRUE (default)}
}
\value{
A data frame
}
\description{
\code{tidy_labs} tidy lab result data
}
\details{
This function takes a data frame with lab results and returns a tidy data
frame. Results will be converted to numeric values and censored data will be
indicated.
}

