% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_data.R
\name{tidy_meds_cont}
\alias{tidy_meds_cont}
\title{Tidy continuous medications}
\usage{
tidy_meds_cont(raw.data, ref.data, sched.data, patients)
}
\arguments{
\item{raw.data}{A data frame with all medications}

\item{ref.data}{A data frame with three columns: name, type, and group}

\item{sched.data}{A data frame with all intermittent medications}
}
\value{
A data frame
}
\description{
\code{tidy_meds_cont} determines which patients have the desired continuous
medications
}
\details{
This function takes a data frame with reference medications or medication
classes and data frames with all continuous and scheduled medications, and
returns a data frame with only the desired medications for each patient. The
data frame passed to ref.data should contain three columns: name, type, and
group. The name column should contain either generic medication names or
medication classes. The type column should specify whether the value in name
is a "class" or "med". The group column should specify whether the medication
is a continous or scheduled medication.
}

