% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_data.R
\name{tidy_services}
\alias{tidy_services}
\title{Tidy services}
\usage{
tidy_services(raw.data)
}
\arguments{
\item{raw.data}{A data frame with service data}
}
\value{
A data frame
}
\description{
\code{tidy_services} tidy medical service history data
}
\details{
This function takes a data frame with medical service history and produces a
tidy version with service begin and end dates/times. It accounts for
incorrect end times from raw EDW data by calculating the end time using the
start time of the next service (unless it was the patient's last service
during the hospitalization). It also combines multiple rows of data when the
patient did not actually leave that service. The data should be read in by
\code{\link[BGTools]{read_edw_data}}.
}

