% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_data.R
\name{tidy_vent_times}
\alias{tidy_vent_times}
\title{Tidy vent times}
\usage{
tidy_vent_times(raw.data, visit.times)
}
\arguments{
\item{raw.data}{A data frame with vent times}

\item{visit.times}{A data frame with discharge date/times}
}
\value{
A data frame
}
\description{
\code{tidy_vent_times} tidy ventilator start/stop data
}
\details{
This function takes a data frame with ventilator times and produces a tidy
version with accurate start and stop dates/times. It accounts for incorrect
end times from raw EDW data. The data should be read in by
\code{\link[BGTools]{read_edw_data}}.
}

