% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{write_docx}
\alias{write_docx}
\title{Write docx to Word}
\usage{
write_docx(mydoc, file.name, add.citation = TRUE, prep = NULL)
}
\arguments{
\item{mydoc}{A docx object}

\item{file.name}{A character string with the name of the project}

\item{add.citation}{An optional logical, if TRUE (default) a citation for R
is added to the document}

\item{prep}{An optional character string with the name of the person who
prepared the data analysis, ignored if add_citation is FALSE}
}
\value{
A docx object
}
\description{
\code{write_docx} save a docx object to a Microsoft Word document
}
\details{
This function takes a docx object and optionally inserts a citation for R,
then writes the docx object to a Microsoft Word document.
}
\seealso{
\code{\link[ReporteRs]{writeDoc}}
}

