% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{analyze_data}
\alias{analyze_data}
\title{Analysis of data}
\usage{
analyze_data(my.data, my.group)
}
\arguments{
\item{my.data}{A data frame with data to be analyzed}

\item{my.group}{A vector to group patients on}
}
\value{
A list
}
\description{
\code{analyze_data} perform data analysis and hypothesis testing
}
\details{
This function takes a data frame with data to be analyzed, performs
hypothesis testing, and returns a list which can be reported in a markdown
document.
}

