% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_data.R
\name{calc_lab_runtime}
\alias{calc_lab_runtime}
\title{Calculate run times for labs}
\usage{
calc_lab_runtime(cont.data, units = "hours")
}
\arguments{
\item{cont.data}{A data frame with serial measurement data}

\item{units}{An optional string with the unit of measure to pass to
\code{\link{difftime}}, defaults to hours}
}
\value{
A data frame
}
\description{
\code{calc_lab_runtime} calculates run time and duration for labs
}
\details{
This function takes a data frame with serial measurement data and produces a
data frame with the duration of time at each value and cumulative run time.
The data frame should have a column called lab.start which is used as the
zero time for calculations.
}

