% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_data.R
\name{calc_perc_time}
\alias{calc_perc_time}
\title{Calculate proportion of time above or below a threshold}
\usage{
calc_perc_time(cont.data, thrshld, meds = TRUE)
}
\arguments{
\item{cont.data}{A data frame with serial measurement data}

\item{thrshld}{A list of the criteria}

\item{meds}{An optional logical indicating whether the data contains serial
medication adminstration or laboratory data, defaults to TRUE}
}
\value{
A data frame
}
\description{
\code{calc_perc_time} calculates percent time above / below a threshold
}
\details{
This function takes a data frame with serial measurement data and produces a
data frame with percent time above or below a threshold for each infusion.
}
\examples{
\dontrun{
calc_perc_time(data, list(~med.rate > 0.4))
# calculates the proportion of time where the rate is above 0.4
}
}

