% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{result_table}
\alias{result_table}
\title{Create a results table}
\usage{
result_table(mydoc, test, table.title)
}
\arguments{
\item{mydoc}{A docx object}

\item{test}{A data frame}

\item{table.title}{A character string}
}
\value{
A docx object
}
\description{
\code{result_table} make a table with results to be inserted into Microsoft
Word
}
\details{
This function takes a data frame and outputs the results of the analysis to a
FlexTable, for inclusion in a Word document.
}
\seealso{
\code{\link[tableone]{CreateTableOne}}
\code{\link[ReporteRs]{FlexTable}}
}

